refactor(tocco-ui): map long as integer
- remove long typeEditable
- remove integer story
- remove decimalScale from number typeEditable

+++ packages/app-extensions/src/formField/defaultMapping.js
@@ -20,7 +20,7 @@ export default {
   'identifier': editableValueFactory('string'),
   'latitude': editableValueFactory('coordinate'),
   'login': editableValueFactory('string'),
-  'long': editableValueFactory('long'),
+  'long': editableValueFactory('integer'),
   'longitude': editableValueFactory('coordinate'),
   'location': editableValueFactory('location'),
   'moneyamount': editableValueFactory('moneyamount'),
